requirements.txt
absl-py==0.10.0
astor==0.8.1
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
docopt==0.6.2
future==0.18.2
gast==0.2.2
glfw==2.0.0
google-pasta==0.2.0
grpcio==1.31.0
gym==0.7.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
isort==4.3.21
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.0.3
mccabe==0.6.1
mujoco-py==0.5.7
num2words==0.5.10
numpy==1.18.5
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyglet==1.3.2
pylint==2.6.0
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.50.2
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
